Refactor simple-tree-core-types
