+++ packages/dds/tree/src/simple-tree/api/index.ts
@@ -125,6 +125,7 @@ export {
 
 export { generateSchemaFromSimpleSchema } from "./schemaFromSimple.js";
 export { toSimpleTreeSchema } from "./viewSchemaToSimpleSchema.js";
+export type { TreeChangeEvents } from "./treeChangeEvents.js";
 
 // Exporting the schema (RecursiveObject) to test that recursive types are working correctly.
 // These are `@internal` so they can't be included in the `InternalClassTreeTypes` due to https://github.com/microsoft/rushstack/issues/3639
