+++ packages/dds/tree/src/simple-tree/api/treeBeta.ts
@@ -7,7 +7,6 @@ import {
 	getKernel,
 	isTreeNode,
 	type NodeKind,
-	type TreeChangeEvents,
 	type TreeNode,
 	type Unhydrated,
 	type WithType,
@@ -15,6 +14,7 @@ import {
 import { treeNodeApi } from "./treeNodeApi.js";
 import { createFromCursor } from "./create.js";
 import type { ImplicitFieldSchema, TreeFieldFromImplicitField } from "../schemaTypes.js";
+import type { TreeChangeEvents } from "./treeChangeEvents.js";
 
 // Tests for this file are grouped with those for treeNodeApi.ts as that is where this functionality will eventually land,
 // and where most of the actual implementation is for much of it.
